Obtain prepared statement config via MFA instead of file path
Instead of specifying the path to a file containing the SQL config
using prepared_statement_source, sqerl now expects a
prepared_statement_mfa key whose value should be a {Module, Fun, Args}
tuple that will return a list of statements when called.

This makes it easier for an application to store the statements in
priv and find them at load time even if said application is included
as a dependency within another application. It also allows
configuration of sqerl prepared statements without using a static
config file.