Refactor tests for pooler as included app
Also moved some db config into the Erlang side. This could go a bit
further by having the Erlang test process shell out for db setup and
cleanup. This way the db won't be left in an inconsistent state when
tests error out and we can get the db config back to being in one
place. It also will make it easier to run the test suite against both
dbs in one go.