Removing spurious calls to COMMIT AND ROLLBACK
sqerl_pgsql_client issued COMMIT and ROLLBACK commands outside
of a transaction. This resulted in harmless warning messages on
the server which are annoying and clutter logs. This commit swaps
those calls for the pg SYNC operation which issues a server-side
checkpoint outside of a transaction which is the correct behavior.