Remove mysql support from sqerl
As the primary maintainer and consumer of sqerl (Opscode), it makes
sense to consolidate our effort supporting what we can commit to
maintaining and testing.

This commit removes support for MySQL, but keeps the modular design to
allow db drivers to be plugged in at start up. This could be useful to
evaluate alternative pg drivers or to bring in support for a different
db type in the future.