Optional ipv6 support
This patch introduces a new optional config key 'enable_ipv6'. It
defaults to false. When true, the 'db_host' config value is parsed
as an ipv6 address/name. If this fails with nxdomain, then the value
is parsed as an ipv4 address/name.

The function inet:getaddr/2 is used to parse the db_host value. As a
result, driver modules will now receive a parsed IP address tuple
rather than a hostname or IP address string (even when enable_ipv6 is
false or unset).