Rework default queries: use first field
The default queries were bogus because they assumed a 'name' field. This
patch uses the first field in the record (really as returned by
RecName:'#info-'/1) as a unique column for use in single row UPDATE,
DELETE, and SELECT (fetch).

The fetch_all and fetch_page queries have been removed from the default
set, but the helper functions for generating these queries are now
exported.

When specifying query names, you can now provide a nested list of
atoms (iolist-esque, but just atoms) which will be joined into a single
atom with 'RecName_' prepended as before. This can be convenient for
programmatic query generation (see implementation of default_queries/1.

Add type export sqerl:atom_list() representing an atom or a deep list of
atoms.