Remove broken datetime input transformer
Epgsql expects a timestamp to come an erlang_time() tuple.  This
transform was converting that to an invalid tagged-tuple of form
{datetime, erlang_time()}.

There was a bug in the transform (using _M for both month and minute) that
 meant it only pattern matched on certain cases (when month == minute) meaning
it appeared as an intermittent error, causing a badmatch in pgsql
pgsql_idatetime:date2j/1.

This commit juts removes the bad transform completely since it's the
wrong thing to do.

change one of the examples so that month==minute in the test to catch regressions of CA-531