Pin back to chef org
Delivery needs the int4range type. A pr has been opened and merged against
epgsql/epgsql. That PR was not merged to master but rather a devel branch
that is currently unreleased.

Until the devel branch is actually released, we need to pin back to the
chef organization epgsql. It should only be a few commits ahead of master
and we will pin back once epgsql releases.