Add new req_timeout parameter to epgsql connection.
This enables (simulated) socket level timeout - tcp can't give it to us,
but erlang can. When req_timeout is set, the epgsql postgres connection
will monitor when it is expected to receive a reply, and terminate the
connectoin with {error, timeout} when it is not received before
(approximately) req_timeout.

Signed-off-by: Marc Paradise <marc@chef.io>
Signed-off-by: Marc Paradise <marc.paradise@gmail.com>