Add verify_ca TLS verify mode
In older versions of Erlang, verify_peer didn't validate hostnames in
many cases. However, since Erlang 19+, it validates hostnames in many
more cases.

While we generally like our hostnames validated, for SQL in
particular, this can be problematic as we are sometimes connected via
an IP that doesn't match the certificate on the server.

This change allows the user to specify `verify_ca` as the verify mode
to mimick the `verify-ca` PG URL option. In this mode, all normal
verification is done, but hostname_check_failed errors are ignored.

Signed-off-by: Steven Danna <steve@chef.io>