Temporarily point to upstream epgsql
Newer versions of epgsql include jsonb support. Because
this is just a PoC effort for now, we'll point directly
to the upstream - otherwise we need to re-do the request-timeout
change [1], which can only be done safely by re-evaluating the current
state of the connection protocol handling and making the change as a new
change.

We'll tackle that work later, if the PoC pans out.

The only other impact here was that epgsql does not support
the req_timeout option in upstream, so I removed it from
sqerl_pgsql_client too.

[1] https://github.com/chef/epgsql-1/commit/7712ee2bbd0eea78025cb7c852ca6944d6001498

Signed-off-by: Marc A. Paradise <marc.paradise@gmail.com>